Edit README.md instructions
It is a bad idea to execute the wizard using sudo since, if done this
way, all commands will be executed as root, which can be dangerous and
even impractical if you want to add instructions that use certain
environment variables.
Also, the caution notice at the beginning of README.md has been removed
as the project is already functional.